microservice_A/requirements.txt
blinker==1.6.2
click==8.1.3
coloredlogs==15.0.1
Flask==2.3.2
humanfriendly==10.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
pydub==0.25.1
PyYAML==6.0
Werkzeug==2.3.6
zipp==3.15.0
